[02.27] Preview Code App.py에서 분리
