+++ controller_flask.py
@@ -7,7 +7,7 @@ import time
 import rclpy
 from cv_bridge import CvBridge
 import cv2
-
+from camera_function.preview_stream import generate_preview, create_ros_subscriber
 
 app = Flask(__name__)
 controller = Controller()
@@ -56,42 +56,14 @@ def raw_callback(x):
     frameTime = endTime
     socketio.emit('preview_image_response', json.dumps(x))
     
-    
-output_frame = None
-lock = threading.Lock()
-def cv_raw_callback(msg):
-    bridge = CvBridge()
-    cv_image = bridge.imgmsg_to_cv2(msg, desired_encoding='passthrough')
-    global output_frame
-    output_frame = cv_image
-
-def create_ros_subscriber():
-    node = controller.con_subscribe_camera_preview(None, cv_raw_callback)
-    # Spin in a separate thread
-    def spin():
-        rclpy.spin(node)
-
-    thread = threading.Thread(target=spin)
-    thread.start()
-
 @app.route("/preview_video_feed")
 def preview_video_feed():
-    return Response(generate(),
+    return Response(generate_preview(),
                     mimetype = "multipart/x-mixed-replace; boundary=frame")
 
-def generate():
-    global output_frame, lock
-    while True:
-        with lock:
-            if output_frame is None:
-                continue
-            (flag, encodedImage) = cv2.imencode(".jpg", output_frame)
-            if not flag:
-                continue
-        yield(b'--frame\r\n' b'Content-Type: image/jpeg\r\n\r\n' + bytearray(encodedImage) + b'\r\n')
 
     
 if __name__ == '__main__':
-    create_ros_subscriber()
+    create_ros_subscriber(controller)
     socketio.run(app, host='0.0.0.0', port=5000, debug=True)
